Simulate database access (aka yada resource properties)
This is an experiment to see if it's possible to check resource
properties, even via a different thread (e.g. a trip to a database),
such that we can determine a 304 Not Modified ahead of time.

Determining status codes ahead of response generation will make it
easier to implement OpenAPI.